+++ ChangeLog.md
@@ -4,6 +4,11 @@ The format is based on
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## v2.13.0 -- YYYY-MM-DD
+
+### Added
+* Filter neighborlists with `freud.locality.FilterSANN`.
+
 ## v2.12.1 -- 2022-12-05
 
 ### Added